Cast result lengths to int before using them as field lengths to avoid warnings when building with Heimdal.

+++ pam_krb5_pass.c
@@ -171,10 +171,10 @@ password_change(struct context *ctx, struct pam_args *args, const char *pass)
             pamk5_error(ctx, "malloc failure: %s", strerror(errno));
         else {
             sprintf(message, "%.*s%s%.*s",
-                    result_code_string.length,
+                    (int) result_code_string.length,
                     (char *) result_code_string.data,
                     result_string.length == 0 ? "" : ": ",
-                    result_string.length, (char *) result_string.data);
+                    (int) result_string.length, (char *) result_string.data);
             krb_pass_utter(ctx->pamh, args->quiet, message);
             free(message);
         }
